Clean up rehome playbook for updating sc cert
The update of the sc-cert will be triggered by cert-mon and
applied by sysinv-agent, applying puppet manifest for that is
not necessary in the subcloud rehoming playbook.

This commit cleans up the unnecessary tasks in the rehoming
playbook, adds a check to ensure the playbook will wait for the
configuration applied before the next step.

Test plan:
Passed - rehoming an AIOSX subcloud with this change, the
subcloud will go online after the playbook. Monitor the sysinv
log of the subcloud, the "platform::config::dc_root_ca::runtime"
and "platform::haproxy::restart::runtime" are applied during
rehoming

Story: 2010230
Task: 46751

Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: Ia4bfc90b616d80955b5ec6e8d8c499ccd2fde7e6