Fix first password change in Ansible for Debian
For Debian, when the user is prompted to change the password on first
login, the strings printed are different from CentOS: Debian prints
"Current password" while CentOS prints "(current) UNIX password". This
difference cause some Ansible tasks that change password to fail. The
same problem was corrected for LDAP in
https://review.opendev.org/c/starlingx/ansible-playbooks/+/863699 .

One of the cases where the password is changed by Ansible tasks is
during bootstrap of a subcloud using Redfish Platform Management
Service.

Test Plan:

PASS: Successfully bootstrap a subcloud using Redfish, checking in the
Ansible log that the first password was successfully changed.

Closes-Bug: 1996196
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: Id3b47224ceb66f834356a37c5746ca1890f6b6bb