Add "/" at the end of path to recognize as directory
Currently /opt/branding is checked with stat.
The patch: https://review.opendev.org/c/starlingx/tools/+/863560
will replace with the symbolic file to /var/branding
This stat is evaluated in when condition later with isdir
but it is not recognized as directory.

The fix is added slash at the end of path, so that
it is recognized as a directory.

Test Plan:
PASS: Place a branding tar in /opt/branding
      before running ansible in fresh install.
      Run ansible and confirm the branding tar file
      is copied to /opt/platform/config/22.12/branding

Partial-bug: 1996182

Signed-off-by: Takamasa Takenaka <takamasa.takenaka@windriver.com>
Change-Id: I3cbc72b8cc3f129768991f5508c2651099641483