Bootstrap failing for ISO based on CentOS
During the bootstrap execution is occurring the failure of it in the
bringup-essential-services playbook.

The commit responsible for that failure is this:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/861222

The fix for it is adding a OS check before run the command inside that
block.

Test plan:
PASS - Install and bootsrap using an ISO based on CentOS
PASS - Install and bootsrap using an ISO based on Debian

Closes-bug: 1995970

Signed-off-by: Samuel Presa Toledo <samuel.presatoledo@windriver.com>
Change-Id: Ie492946c95f27ce98a195b4cdb0c00e6df2e7665