Update packages prestage logic for Debian
Changes have been made which accomplish prestaging of
the repo and container images on the Debian subcloud. The existing
prestage logic has been modified so that the kickstart can find the
prestaged software packages for debian.

In this commit:
  - software packages are now prestaged under /opt/platform-backup/ostree_repo.
  - registry user images prestage location remains unchanged (versioned).
  - remove code related to prestage_packages_completed flag as
    it is no longer relevant. The ostree repo prestage can be re-run.

Test Plan:
PASS: Verify that the ostree_repo is prestaged at
      /opt/platform-backup.

PASS: Verify that the container images are prestaged at
      /opt/platform-backup/<release id>.

PASS: Verify that remote install uses prestaged ostree_repo if exists.

PASS: Verify that reinstall uses prestaged ostree_repo if it exists.

Story: 2010120
Task: 46840

Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: Ia4aeb577b71b407ffd71bc41b8cd14b7fa0301e9