Remove failed_when condition for branding restore
Details: Remove failed_when condition for branding restore

This condition was added in order to unblock SX upgrade work. The
original directory now in Debian is managed by ostree and not writable.
After dependency CR is merged, the branding directory is relocated to
a writable place. This condition is no longer needed.

Depends-On:  https://review.opendev.org/c/starlingx/tools/+/863560

Test Plan:

PASS: SX upgrade and branding is restored

Task: 46298
Story: 2009303
Signed-off-by: Junfeng (Shawn) Li <junfeng.li@windriver.com>
Change-Id: I47ee72fb2ee549c609a98d99c22c3ac110fa7004