Synchronize patching metadata on remote install
Handle case where subcloud redfish installation occurs via a patched
system controller.

In the problem case, the system controller is running the patch ostree
commit revision. During the subcloud miniboot kickstart, the subcloud
pulls down and boots into the same ostree patch commit.  However, if
patch metadata is not propagated to the subcloud, the subcloud sw-patch
tooling is not correctly aligned with the ostree commit/patch-level
running on the subcloud.

This commit adds a patch metadata synchronization block to the
subcloud install ansible code.

The patch synchronization first checks if the system controller and
subcloud have:
  1) Different reported patch levels from 'sw-patch query', and
  2) The same ostree revision.

If the system controller and subcloud:
- Are at the same ostree commit and have the same sw-patch output:
    - No action is required. We skip the metadata synchronization.
- Are at the same ostree commit but have different sw-patch output:
    - The subcloud is running at the same patch level but has
      out-of-date patch metadata, and we proceed with the /opt/patching
      sync.
- Are not at the same ostree commit:
    - The subcloud has been installed via an ISO (non-redfish), and will
      need to be patched post-install. We therefore skip the metadata
      synchronization.

The install.yml playbook is run in the following cases:
- add subcloud (NOT for rehome)
- reinstall
- restore (if 'with_install' is specified in payload)
- upgrade

LIMITATION:

This commit does not synchronize patching state between system
controller and subcloud in the case where a patched system controller is
used to install a subcloud using a *non-patched ISO*. In this case,
the subcloud does not pull from the system controller ostree repo, and
therefore doesn't obtain the patched ostree commit(s).
The subcloud will be required to be patched after bootstrapping.
We will handle this situation as required in a new commit.

TEST PLAN:

PASS:
- Tests with patched system controller:
  - Add simplex subcloud via redfish (miniboot).
      - Verify that subcloud boots with patched ostree commits, and that
	/opt/patching metadata is synchronized to subcloud.
      - Upon subcloud unlock, verify that sw-patch query on subcloud
	aligns with the ostree commits.
  - Add simplex subcloud booted from ISO (non-redfish). Verify that this
    patching synchronization is skipped - i.e., subcloud boots
    with non-patched ostree commits, and that /opt/patching is empty on
    subcloud.
  - In the duplex case it has been confirmed that patches are propagated
    to controller-1 when controller-0 has been patched in this manner.
PASS (regression):
- Tests with unpatched system controller:
  - Verify add subcloud via redfish. Verify that this patching
    synchronization is skipped.

Closes-Bug: 1995995

Change-Id: Ifdc361280dec2f19c54a1303fce6a46eb66cc906
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>