Add support for DC admin network
This commit allows a subcloud to be provisioned with an
optional admin network, which will take precedence over the
existing management network for communication between the
subcloud and system controller.

The management network will still exist on the subcloud, but
will be a private network.

Because the management network and it's parameters are
embedded in many parts of the system, having a separate
admin network makes it much easier to change the parameters
of this network (subnet, gateway, etc) after a subcloud has
been provisioned.

This initial commit allows one to provision an admin network
on subcloud add (bootstrap).  Further work will be needed
to support other operations such as rehoming and subcloud
update.

Test Plan:

- Bootstrap and install non-DC AIO-SX with admin network defined.
  PASS: No impact.
  PASS: Admin network is not created.

- Bootstrap and install DC subcloud with admin network defined.
  PASS: Ensure sc-adminep-certificate has the admin IP floating
        address.
  PASS: Ensure admin network is created.

- Manually allowing admin network to be created on non subcloud
  system
  PASS: Ensure an ansible re-run is successful.
  PASS: Ensure adding the admin network after the 1st run
        allows admin network to be created on the re-run

Regression:

- Bootstrap and install non-DC AIO-SX system with no admin network
  defined
  PASS: No impact.

- Bootstrap and install DC subcloud with no admin network defined.
  PASS: Ensure sc-adminep-certificate has the mgmt IP floating
        address.
  PASS: Ensure the subcloud can become online and in-sync with the
        management network (no impact).

Failure Path:

PASS: Verify operation of admin subnet address range validations.

Story: 2010319
Task: 46909
Depends-On: https://review.opendev.org/c/starlingx/config/+/863033

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: I6fee5bd0c628380f4edcc41fb66f2a601999830f