Remove auto-add of ghcr.io Docker registry
The code related to containerization components was added in
previous releases to enable upgrades. Some of this code is no
longer needed and can be removed to reduce code clutter.

Remove auto-add of ghcr.io Docker registry in the ansible-playbooks
git repo which is no longer required.

Test Plan:
PASS: Deployed 22.06 CentOS release in AIO-SX and AIO-DX.
      Completed the platform upgrade from 22.06 to 22.12 using
      the ISO generated with the change.

Story: 2010445
Task: 46903

Signed-off-by: Ramesh Kumar Sivanandam <rameshkumar.sivanandam@windriver.com>
Change-Id: Idedc98f0535d3e77c47143cd900204133b6d3591