Restore hung at ceph-osds mount
Restore process was not completing properly due to
command lvcreate be prompting a message to the user
and hung expecting for a confirmation.
The prompt message returned by lvcreate was:
WARNING: ext4 signature detected on /dev/cgts-vg/ceph-mon-lv
  at offset 1080. Wipe it? [y/n]

This update adds an auto confirmation to the command
and sends a kill signal instead of the default sigterm
which was not terminating the lvcreate proccess.

Test Plan:
PASS - Backup and restore in AIO-SX lab

Closes-bug: 1998019

Signed-off-by: Daian Cardoso Sganderlla <Daian.CardosoSganderlla@windriver.com>
Change-Id: I1fa74ca0ae1974caa73df11a546890cab8520ca0