Fix prestage failure due to incorrect ostree remote
When we try to prestage the subcloud, ostree_repo is transferred
to /opt/platform-backup on the subcloud.

However, if ostree_repo had already been prestaged there with a
prestage iso, then this will fail. This is because the remote for
the previously prestaged ostree repo will wrongly point to
/instboot/ostree_repo, and the install fails.

This commit rectifies the issue by deleting the remote first and
then adding the new remote.

Test Plan:
PASS: Verify that subcloud prestage succeeds in the presence of
      a previously prestaged ostree_repo.

Closes-Bug: 1998250

Change-Id: Ie1169e29966007186497ec5b5c8cfb2819786ddc
Signed-off-by: ssharma3 <shrikumar.sharma@windriver.com>