Fix default gateway detection in optimized restore
Additional follow up for https://review.opendev.org/c/starlingx/ansible-playbooks/+/866403.

This change is required for system that required the previous review.
The default routes were not being detected so a change in how the are
read was required.

TEST PLAN
PASS: Perform backup and optimized restore on afflicted system (AIO-SX)

Closes-Bug: 1998504
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: Ie9aa9261f4e0b80cfff977a84802e65ca60442f2