Add rehoming support for DC subcloud admin network
The admin network is an optional network that a user can choose
to deploy to handle communication between a distributed cloud
subcloud and system controller.  Because the management network
is so embedded into the operations of a StarlingX system, the
introduction of the admin network allows a subcloud operator to
more easily change the parameters of this network without
affecting the operation of the subcloud itself.  For example,
if routing changes occur in any upstream routers, a subcloud
using a management network would have to be re-installed.  A
subcloud using the admin network will be able to be modified
to handle the physical networking changes without affecting
the normal operation of the subcloud management operations.

If an admin network is defined in the bootstrap parameters of
the subcloud, it will take precedence over the management
network for communication with the system controller.

This commit introduces changes need to support changes in the
admin network in the case that a subcloud is re-homed to a new
system controller.  As part of these operations, a user can
define new parameters for the admin network to account for any
changes on the physical network.

Testing:

AIO-SX: A subcloud using the admin network can be rehomed
        to a new system controller using the existing
	admin network parameters.

AIO-SX: A subcloud using the admin network can be rehomed
        to a new system controller using new admin network
	parameters.

AIO-DX: A subcloud using the admin network can be rehomed
        to a new system controller using the existing admin
        network parameters.

AIO-DX: A subcloud using the admin network can be rehomed
        to a new system controller using new admin network
	parameters.

AIO-SX: (Regression) Update a subcloud initially provisioned
        with the management network to use a newly created
	admin network.

AIO-SX: (Regression) Update a subcloud with new admin network
        subnetting.

AIO-SX: (Regression) A subcloud using the management network
        for communication with the system controller can
	be rehomed to a new system controller.

Story: 2010319
Task: 47044

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: I53bda7d35e3d826be799a9faaa5968eeaee97912