Fix backup failure due the ldap conf file missing
When we do the backup, the process tries to convert the ldap config
directory into a LDIF file.

However, if the ldap config directory doesn't have the cn=config.ldif
file, the backup process fails. Nevertheless, this file may not exist
on a subcloud.

This commit sets backup of ldap config directory as conditional, to
only execute if the necessary files exist, otherwise it must follow the
normal flow.

Test Plan:
PASS: The backup and restore process without the ldap config directory
in AIO-SX.
PASS: The backup and restore process without the ldap config directory
in AIO-SX subcloud.
PASS: The backup and optimized restore process without the ldap config
directory in AIO-SX subcloud.

Closes-Bug: 1999151

Signed-off-by: Guilherme Schons <guilherme.dossantosschons@windriver.com>
Change-Id: I0244683f6373a4e9edc789caa5c666a9e1800b19