Increasing retries and delays in restore-ceph-data playbook role
When running the restore-ceph-data playbook roles, one of the tasks
is "Bring up ceph OSDs", which retries 5 times and delays 3 seconds
between retries. Although it can't be enough in some cases.

We increased the delay between retries to 10 seconds and the
number of retries to 6 trying to avoid this issue.

Test Plan:
PASS: Upgrade successfully from STX7.0 Simplex to STX8.0

Closes-Bug: 1999168

Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>
Change-Id: I16f40a040a96215ea57055e6e661e09a66504522