Fix two defects in apply-rvmc-job playbook role
1. The kubectl command that was used to wait for the ready state is not
reliable. Sometimes the pod completed its job but k8s did not detect its
Ready condition. Either the pod never changed to Ready condition or k8s
might have missed the detection of condition transition to Ready.
Therefore, a rescue block is added to double confirm that the pod indeed
failed before failing the playbook.

2. The failed condition for "Ensure rvmc namespace is created" task
was wrong. The "AlreadyExists" keyword is not quoted in the error message.

Test plan:
1. Verify successful batch subcloud deployment in a scale lab.
2. Updating install values of a subcloud with wrong bmc address.
   Verify that the install playbook fails as expected and pod logs with
   correct failure reason is captured in ansible log.

Closes-Bug: 1999811

Signed-off-by: Li Zhu <li.zhu@windriver.com>
Change-Id: Ib3ede454db5daa9f1585d4d9892e1b3f4f57df1c