Stop multus pod from requesting CPU resources
This change will not allow platform pods: i.e. multus,
from requesting CPU resources.

Due to limitations in the kubernetes customizations
around resource tracking and platform CPUs. If the
platform pod uses up a portion of the CPUs,
it may limit the ability of the application pods to
utilize all of the designated CPUs.

Testing:
 - Execute command: kubectl describe node
   -> Name:                                     CPU Requests
      kube-multus-ds-amd64-dccsj                   0 (0%)

Closes-Bug: 2000177

Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>
Change-Id: I69017f4c631f2bea895c2d878ccf615143faad7e