Fix calico bring-up failure in optimized restore
On certain system, the networking service can be restarted.
On systems were it is possible it should be restored that way.
Otherwise, fallback to the original method of restoring
networking by using ifup/ifdown manually.

It was seen that if restarting networking service is possible
but is not done,  Calico cni will fail to be created when Kubernetes
is brought up later in the playbook.

This only impacts certain systems.

TEST PLAN
PASS: Run optimized restore with --registry-images (AIO-SX)
  * Run on affected and unaffected systems
PASS: Run optimized restore without --registry-images (AIO-SX)
  * Run on affected and unaffected systems

Closes-Bug: 2000184
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: I97b9fc2bbbe47ed93edbdae3f4b02b6243181677