Backup delete fails if backup path doesn't exist
Adds a check to verify that the subcloud backup location exists before
attempting to delete it. If the directory does not exist, the
operation will now fail and return an error message to the user.

Test Plan:
1. PASS - Validate that the backup delete with --local-only option
will fail if the backup folder does not exist locally on the subcloud.
2. PASS - Validate that the backup delete will fail if the backup
folder does not exist in the central System Controller.
3. PASS - Validate that the backup delete with --local-only option
 will succeed if the backup directory exists locally on the subcloud.
4. PASS - Validate that the backup delete will succeed if the backup
directory exists in the central System Controller.
5. PASS - Validate, on all cases above, that the backup datetime and
backup status are only reset when the delete operation succeeds in
finding and deleting the backup directory.

Closes-Bug: 2000320

Change-Id: I90b1e9b3df6b30aa0a37a7379d8a2e80d3c37974
Signed-off-by: Andre Carneiro <Andre.DexheimerCarneiro@windriver.com>
Signed-off-by: Guilherme Costa <guilherme.costa@windriver.com>