Fix deletion of ostree repo under /var/run after prestage
When prestaging for upgrade, the ostree repo was first copied
to /var/run prior to rsyncing to the subcloud. After rsync, the
repo in /var/run was not deleted, causing collect logs from this
directory to include the ostree repo as well.

This commit addresses this issue by avoiding the use of /var/run
and rsyncing directly from
/var/www/pages/feed/rel-<rel_id>/ostree_repo. Since we are now
rsyncing directly from the feed, we no longer need to lock down
the repo, thereby ensuring that any new patches in the feed are
prestaged in future runs.

Test Plan:
PASS: Verify that prestage for upgrade uses the repo from the
      feed directory and does not leave any artifacts in the
      /var/run directory.

PASS: Verify that prestage for upgrade executes without any
      regression.

PASS: Verify that prestage for reinstall executes without any
      regression.

Closes-Bug: 2003271

Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: I10a8bad066cd5761f3c86859a2ff9c3ef90d31e7