Fix backup-restore with expired certificate
This commit updates the certificates database table according to
the ssl_ca certificates installed during the restore procedure.
The installed certificates can be obtained either from the
backup file or by the file specified by the parameter
'ssl_ca_certificate_file'.
This parameter was implemented on the following reviews:

https://review.opendev.org/c/starlingx/ansible-playbooks/+/852785
https://review.opendev.org/c/starlingx/ansible-playbooks/+/855362

In the end of the original restore procedure, the certificates
database table is restored according to its state from the
backup file, instead of considering the ssl_ca certificates
installed earlier in the restore playbook from the file
specified in the ssl_ca_certificate_file parameter.
This can cause a mismatch between the certificate database table
and the certificates actually installed in the filesystem,
making the database system commands such as
'system certificate-uninstall', not work properly.

The solution was to backup the certificate database table to
a tmp file before postgres DB restore and, once postgres DB is
restored, replace the certificate table in postgres with the
data from the tmp file. To guarantee that the certificates
are inserted correctly to the database table, a sequence
reset step was created, this step rearrenge the table rows
avoiding a duplicate row id error. That will ensure the
database table matches with certificates in the filesystem.

PASS: Tested on a SX VM with an expired certificate,
using the ssl_ca_certificate flag and without using it.

PASS: Tested the commands to install
(system certificate-install -m ssl_ca <cert.pem>)
and uninstall certificates
(system certificate-uninstall -m ssl_ca <ssl_uuid>).

PASS: Tested scenario to restore backup with multiple
expired certificates.

PASS: Tested scenario adding a valid certificate
before running backup process.

Closes-Bug: 1984116

Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>
Change-Id: Iaac2cfe9f456e4ae4a5ba21a6cf8ca96ca1e1eb8