Remove ssl-cert-snakeoil.key from backup file
An error was found trying to restore a backup from a system with
comitted patches. The error is about PostgreSQL trying to use the
ssl-cert-snakeoil.key certificate, but the value from this certificate
mismatches the expected value.

This certificate is changed when a patch is applied.

By removing this file from the backup, the restore process works fine,
with the system generating a new certificate

Test Plan:
- PASS: Perform B&R AIO-SX with patches applied.

- PASS: Perform B&R AIO-SX Without patches applied.

- PASS: Perform B&R Standard with patches applied.

- PASS: Perform B&R Standard without patches applied.

Closes-Bug: 2004189
Signed-off-by: Guilherme Schons <guilherme.dossantosschons@windriver.com>
Change-Id: Ic1e2c18f3f7f68687531e1b471b4dfda4738ff2f