Fix CommonName usage in cert migration playbook
In migrate-platform-certificates-to-certmanager playbook, the
'subject_CN' field of the provided inventory file is not being used to
change the CommonName (CN) for the migrated certs. The playbook uses
'system-openldap' for the OpenLDAP local cert and the OAM IP for the
other certs.

Changed the playbook's template to fix this, defaulting to the
older values if the user doesn't provide the 'subject_CN' field.

The certificates affected by the playbook (and this change) are:
'system-registry-local-certificate', 'system-restapi-gui-certificate',
'oidc-auth-apps-certificate' and 'system-openldap-local-certificate'.

Test Plan:
PASS: Follow the steps in
      https://docs.starlingx.io/security/kubernetes/migrate-platform-certificates-to-use-cert-manager-c0b1727e4e5d.html
      to execute the cert manager migration playbook. Provide the
      'subject_CN' field in the inventory file.
      Check each of the certificates affected by the playbook (listed
      above).
      Observe that CN field is set to the provided 'subject_CN'
      value.

PASS: Follow the steps in
      https://docs.starlingx.io/security/kubernetes/migrate-platform-certificates-to-use-cert-manager-c0b1727e4e5d.html
      to execute the cert manager migration playbook. Don't insert the
      'subject_CN' field in the inventory file.
      Check each of the certificates affected by the playbook (listed
      above).
      Observe that the CN field is set to 'system-openldap' for the
      'system-openldap-local-certificate' and  OAM IP for the others.

Closes-Bug: 2007686.

Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>
Change-Id: I75c6de58ef9ee88effaa9d973de685be42648590