Delete certs spec file after migration playbook
In migrate-platform-certificates-to-certmanager playbook, a file is
created to store certificate spec before applying it to cert-manager.
This file has sensible data, and should not be maintained in the
system after the usage.

File ownership will be 'root' and permissions '0640' to limit access
while it is still in the system. File will be removed after usage in
the playbook.

Test Plan:
PASS: Follow the steps in
      https://docs.starlingx.io/security/kubernetes/migrate-platform-certificates-to-use-cert-manager-c0b1727e4e5d.html
      to execute the cert manager migration playbook.
      Wait for the playbook to finish.
      List the files for the folder '/tmp'.
      Verify that the file 'platform_certificates.yaml' is absent.

Closes-Bug: 2008250

Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>
Change-Id: I899bf6920fe2e50079fd1dafc922bae22da47b6c