Update rehome subcloud playbook to update rootca cert
In this commit, the rehoming playbook is updated
to ensure the subcloud will have the same k8s
rootca cert as the new central cloud
post rehoming.

Test Plan:
PASS: Rehome a subcloud.
dcmanager subcloud add --migrate
--bootstrap-address <controller-0-oam-address>
--bootstrap-values <bootstrap_values>
--sysadmin-password <subcloud password>
PASS: open the /etc/kubernetes/pki/ca.crt of the
subcloud and verify that the cert is
the same cert of the central controller.
PASS: create a strategy
- sw-manager fw-update-strategy create
- sw-manager kube-rootca-update-strategy creat
- sw-manager kube-upgrade-strategy create
- sw-manager patch-strategy create
- sw-manager upgrade-strategy create
then rehome a subcloud and verify that the subcloud
has the same k8s rootca cert as the central
controller.

Closes-Bug: 2009827

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: Ia116dcbcd1f02c6b3d56a0529e16617853ba3b08