FPGA plugin cleanup
As part of this commit, all files related to the FPGA plugin are removed
and cleaned up.
FPGA plugin was added in stx 3.0 and is no longer used in stx 9.0.

TEST PLAN:

PASS: building and bootstrap is successful
PASS: intel-fpga-plugin image is not present in the image list
      which is checked using command "crictl images | grep
      'intel-fpga-plugin'"
PASS: intelfpga=enabled is not mentioned under k8s_plugin in
      default.yml

Story: 2010604
Task: 47911

Signed-off-by: Md Irshad Sheikh <mdirshad.sheikh@windriver.com>
Change-Id: Icbbed5e8c303a9f6013e6454dc2fae3f3eb24e7b