Wait for online at end of optimized restore
It was observed that in rare cases, if there is
any instability of the management network that the
system may not be in on online state by the end of the
optimized restore playbook.

This change will attempt to mitigate such a scenario
by adding a retry loop when checking if the system is online.

TEST PLAN
PASS: Optimized restore is successful (SX)

Story: 2010117
Task: 47646
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>
Change-Id: I9a11145fe18e802b1739b7c2ca1bd7c1b02b2e07