Change subcloud backup transfer to synchronize
If the user makes a backup of a subcloud or any other operation that
requires ssh/scp as root by hostname using the System Controller,
the subcloud host key will be saved on the root known_hosts. If the
subcloud host key changes, like if the system is reinstalled, and
another backup is attempted, the operation will fail because of the
presence of the previous host key on the System Controller root
known_hosts.
The backup transfer method was changed from running rsync using command
to the synchronize module from Ansible posix collection, so the ssh
handling is taken care by Ansible.

Test plan:
  PASS:
  - Backup subcloud1 using dcmanager
  - Change subcloud1 host keys
    (by removing old keys and running "ssh-keygen -A")
  - Backup new subcloud and verify that the process ended without errors
  PASS:
  - Backup subcloud1 using dcmanager
  - Delete and reinstall subcloud1
  - Backup new subcloud and verify that the process ended without errors

Closes-Bug: 2011604

Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>
Change-Id: I561b496be978a13fb2bdded0acf6f6534aed42ac