Update ceph wipe in multipath
When attempting to execute the "wipe" command on
multipath disks, warnings are generated by parted,
which causes the script to return an error code of 1.
To address this issue, the script now includes
the use of "kpartx" to update the disk partitions,
which effectively avoids the parted error and allows
for successful execution of the "wipe" command on
multipath disks.

Test plan
  PASS: Mulitpath simplex restore
        with wipe_ceph_osds=true
  PASS: Without multipath simplex restore
        with wipe_ceph_osds=true

Closes-bug: 2011736
Signed-off-by: Lucas Borges <lucas.borges@windriver.com>
Change-Id: I609a63c00e027d49ee92724c620b71bb71b0ddc5