Fix cert-manager tags for stx6/stx7 during AIO-SX upgrades
The tags for application cert-manager differ when upgrading
from stx6/stx7 to stx8. When FROM is stx6, cert-manager uses
tags v0.15.0, and when FROM is stx7 the tags are v1.7.1.

This commit adds a logic to select the right tags for each
release during an AIO-SX upgrade. To enable this, it reverts
the tags changed on commit [1].

[1] https://review.opendev.org/c/starlingx/ansible-playbooks/+/869572/10/playbookconfig/src/playbooks/roles/common/load-images-information/vars/k8s-v1.21.8/system-images.yml

Test Plan
PASS: stx6 -> stx8 AIO-SX upgrade
PASS: stx7 -> stx8 AIO-SX upgrade

Closes-bug: 2016352

Change-Id: I5d3c511a15705a545ea74229249fe1e3864cd506
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>