Backup and restore subcloud alarms
Currently the fm alarm table is erased over
system backup as the expectation is for the
components that raised these alarms to re-raise
them upon system restore. However, the 280.xxx
alarms are only raised and cleared when the
subcloud state changes from online to offline
and offline to online respectively.
If a subcloud is offline at the time the system
controller backup is taken and remains offline
when the system controller is restored from its backup;
the subcloud is offline alarm is missing.

The most resources effective and also simplest solution
is to backup and restore alarms raised by DC audit.

Test Plan:
PASS: Backup a system controller while a subcloud
is offline then restore the system controller
from the backup while the subcloud is still offline.
Verify that 280.001 alarm still exists for the subcloud.

PASS: Backup a system controller while a subcloud is
offline then restore the system controller from the
backup while the subcloud is back online. Verify that
there's no 280.001 alarm for the subcloud.

PASS: Backup a system controller when there is no
subcloud offline then restore the system controller
from the backup.

Closes-Bug: 2018005
Change-Id: I4b3f7d26fa8935d5ffc9aaa324b5334eeaa4f15b
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>