Fix ldap restore failure when backup taken from controller-1
The issue is, in a Duplex system when backup is taken from
controller-1, the restore fails due to SLAPD failure.

SLAPD fails to start because, the backup file
/etc/ldap/schema/cn=config/olcDatabase={1}mdb.ldif defines
the value, provider=ldaps://controller-0 in the
replication-sync(olcSyncrepl) parameter(directive).

When we do restore from controller-0 the provider ID
controller-1 is expected.

NOTE: Restore has to be from controller-0 as for the restore
operation, controller-0 is implicit. Since restore would be
used in disaster recovery/install scenario on duplex, restore
is expected on controller-0.

As a solution, conversion of provider value in sync-replication
directive(olcsyncrepl) is made to point to controller-1.

In AIO-SX, controller-0 provider is set to controller-1. This is
the default value for AIO-SX. The changes mentioned in this review
does not change the value for AIO-SX Backup and Restore.

TEST PLAN
PASS: AIO-SX: Backup and Restore passed
PASS: AIO-DX: Backup and Restore passed

Closes-Bug: 2017927

Change-Id: Ib6c7655e8a8e0fbe8d1426918839054de44ebe64
Signed-off-by: ksingh <kirti.singh@windriver.com>