Adjust Ceph monmap during system restore
When restoring from a backup or doing AIO-SX upgrades, the
Ceph monitor may not get up running after restoring data
from the OSDs.

This happens because sometimes ceph mon was not able to
form quorum trying to connect to the default 'mon.a'
monitor created automatically on first run. In this case
the monitor name must be 'mon.controller' for AIO-DX and
'controller-0' for other configurations.

To fix this scenario the monmap is extracted, configured
and injected back. Then the ceph monitor can start
successfully.

Test-Plan:
  PASS: AIO-SX B&R ipv4 Optimized
  PASS: AIO-SX B&R ipv4 Legacy
  PASS: AIO-SX B&R ipv6 Optimized
  PASS: AIO-SX B&R ipv6 Legacy
  PASS: AIO-SX Upgrade ipv4 stx7.0 to stx8.0
  PASS: AIO-SX Upgrade ipv6 stx7.0 to stx8.0
  PASS: AIO-DX B&R ipv4

Closes-bug: 2018518

Change-Id: I249584f566536f804b8423fe33b8f54f571e5bbd
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>