Set tar --wildcards flag for restore-openstack
After the Debian migration it was verified that the --wildcards flag was
necessary in order to use the regular expressions with "tar".
This change updates the command line accordingly.

Test Plan:
PASS - Build package playbookconfig
PASS - Build ISO including new package
PASS - Deploy an AIO-SX system with ISO
PASS - Verify that restore-openstack was updated
PASS - Run an ansible-backup
PASS - Run restore_openstack.yml successfully *
* This playbook will not successfully finish while the
fix [1] is not merged, but the test was able to pass the MariaDB
step that was broken and went further on the execution.

Story: 2010560
Task: 47737

[1] https://review.opendev.org/c/starlingx/openstack-armada-app/+/876598

Signed-off-by: Romulo Leite <romulo.leite@windriver.com>
Change-Id: Ic81262ef36e5c83c05d056260649a445b27dbffc