Fix subcloud mgmt to admin update on AIO-DX
Currently there's a bug that prevents a user from updating a
subcloud to use the admin network from a previously used
management network on an AIO-DX system.

The subcloud update playbook will fail when updating controller-1
while trying to determine the interface used for communication
with the system controller.

The cause of this bug is simply an unquoted ansible variable:
target_net.

In addition, this bug fix moves the setting of the controller-1
interface fact to be after the check for whether such an
interface is assigned.

This makes the processing of the facts for controller-1 the same
as that for controller-0.

Testing:

- Update an AIO-DX from management network to admin network

Story: 2010319
Task: 47707

Signed-off-by: Steven Webster <steven.webster@windriver.com>
Change-Id: I38a307f214ae7ea01d41330ba368968e6ed2f1f9