Bug-fix: preserve ownership and permission of keystone.conf
During the subcloud rehoming, the keystone.conf is backed up and
restored by the ansible user with root privilege, which changes
its ownership from keystone to root, and results in the subcloud
cloud restore fail. This commit preserves the ownership and
permission when restoring the keystone.conf during rehoming.

Test plan:
1. Rehome a subcloud with this change -- passed.
2. Verify the ownership and permission of the keystone.conf under
/etc/keystone/keystone.conf not changed after rehoming -- passed.
3. Backup and restore the subcloud after rehoming -- passed.
4. Restore subcloud with "restore_mode: old" -- passed.

Closes-bug: 2018938
Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: Ic4547500433bae6ee8739e406b741f04c714cea4